+++ src/main/kotlin/blackjack/domain/PlayerResult.kt
@@ -1,3 +1,5 @@
 package blackjack.domain
 
+import blackjack.domain.participant.Player
+
 data class PlayerResult(val player: Player, val status: GameResultStatus)
