+++ src/test/kotlin/blackjack/DealerTest.kt
@@ -1,7 +1,7 @@
 package blackjack
 
 import blackjack.domain.Card
-import blackjack.domain.Dealer
+import blackjack.domain.participant.Dealer
 import blackjack.domain.Rank
 import blackjack.domain.Suit
 import org.assertj.core.api.Assertions.assertThat
