+++ src/test/kotlin/blackjack/PlayerTest.kt
@@ -1,7 +1,7 @@
 package blackjack
 
 import blackjack.domain.Card
-import blackjack.domain.Player
+import blackjack.domain.participant.Player
 import blackjack.domain.Rank
 import blackjack.domain.Suit
 import org.assertj.core.api.Assertions.assertThat
